% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.recipe.R
\name{read.recipe}
\alias{read.recipe}
\title{Read recipe from yaml file}
\usage{
read.recipe(x)
}
\description{
This function reads a recipe described in a yaml file. Fields must contain:
\itemize{
\item title: the title of the recipe
\item date: the date the recipe was written
\item author: the author of the recipe
\item email: the email of the author
\item ingredients: a list of ingredients
\item description: a list of instructions
}
}
\examples{
template <- system.file("recipes/template.yml", package="eatR")
read.recipe(template)

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}

