% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.recipe.R
\name{add.recipe}
\alias{add.recipe}
\title{Add recipe from yaml file}
\usage{
add.recipe(x = "new.recipe.yml", title = NULL, date = NULL,
  author = NULL, ingredients = NULL, description = NULL)
}
\description{
This function reads a recipe described in a yaml file. Fields must contain:
\itemize{
\item title: the title of the recipe
\item date: the date the recipe was written
\item author: the author of the recipe
\item email: the email of the author
\item ingredients: a list of ingredients
\item description: a list of instructions
}
}
\examples{
add.recipe("test")
file.show("test.yml")

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}

